Separate sockets state and more route refactoring
